Some cleanup / refactoring
