+++ site/src/components/index.tsx
@@ -2,3 +2,4 @@ export * from "./Button"
 export * from "./Confetti"
 export * from "./EmptyState"
 export * from "./Page"
+export * from "./Triangle"
