+++ site/src/pages/Workspaces.tsx
@@ -1,45 +1,10 @@
 import React, { useState } from "react"
+import { Dialog, DialogActions, Button, DialogTitle, DialogContent, makeStyles, Box, Paper } from "@material-ui/core"
 import { AddToQueue as AddWorkspaceIcon } from "@material-ui/icons"
 
-import { Confetti, EmptyState, Page, SplitButton } from "./../components"
+import { EmptyState, Page, SplitButton } from "./../components"
 
-import { Dialog, DialogActions, Button, DialogTitle, DialogContent, makeStyles, Box, Paper } from "@material-ui/core"
-
-const useStyles2 = makeStyles((theme) => ({
-  "@keyframes spin": {
-    from: {
-      transform: "rotateY(0deg)",
-    },
-    to: {
-      transform: "rotateY(180deg)",
-    },
-  },
-  triangle: {
-    animation: "$spin 1s ease-in-out infinite alternate both",
-  },
-}))
-
-export const Triangle: React.FC = () => {
-  const size = 100
-
-  const styles = useStyles2()
-
-  const transparent = `${size}px solid transparent`
-  const colored = `${size / 0.86}px solid black`
-
-  return (
-    <div
-      className={styles.triangle}
-      style={{
-        width: 0,
-        height: 0,
-        borderLeft: transparent,
-        borderRight: transparent,
-        borderBottom: colored,
-      }}
-    />
-  )
-}
+import { CreateWorkspace } from "./CreateWorkspace"
 
 export const Workspaces: React.FC = () => {
   const styles = useStyles()
@@ -84,28 +49,23 @@ export const Workspaces: React.FC = () => {
           <EmptyState message="No workspaces available." button={button} />
         </Box>
       </Paper>
-      <Dialog fullWidth={true} maxWidth={"lg"} open={open} onClose={handleClose}>
-        <DialogTitle>New Workspace</DialogTitle>
-        <DialogContent>
-          <Confetti
-            style={{
-              minHeight: "500px",
-              display: "flex",
-              flexDirection: "column",
-              justifyContent: "center",
-              alignItems: "center",
-            }}
-          >
-            <Triangle />
 
-            <Box m={"2em"}>NEXT STEP: Let's create a workspace with a v2 provisioner!</Box>
-          </Confetti>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleClose}>Close</Button>
-        </DialogActions>
-      </Dialog>
+      <CreateDialog open={open} handleClose={handleClose} />
     </Page>
+  )
+}
+
+const CreateDialog: React.FC<{ open: boolean; handleClose: () => void }> = ({ open, handleClose }) => {
+  return (
+    <Dialog fullWidth={true} maxWidth={"lg"} open={open} onClose={handleClose}>
+      <DialogTitle>New Workspace</DialogTitle>
+      <DialogContent>
+        <CreateWorkspace />
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={handleClose}>Close</Button>
+      </DialogActions>
+    </Dialog>
   )
 }
